Allow everyone with write permission to push to main (#36)
This allows everyone with write permission (i.e. everyone in our organization) to write to the main branch. Changes still need to be first pushed to a unprotected branch and reviewed in a pull request, but clicking on the merge button is now possible for everyone aside the admin/maintainer team.

Signed-off-by: Eduard Itrich <eduard@itrich.net>